+++ test/statistics_base_test.dart
@@ -88,6 +88,18 @@ void main() {
       expect(parseDateTime('50/8/30 12.11.10 -0300X'), isNull);
     });
 
+    test('cast', () {
+      num n = 123;
+
+      expect(n.cast<int>(), allOf(isA<int>(), equals(123)));
+      expect(n.cast<double>(), allOf(isA<double>(), equals(123)));
+
+      num d = 3.14;
+
+      expect(d.cast<int>(), allOf(isA<int>(), equals(3)));
+      expect(d.cast<double>(), allOf(isA<double>(), equals(3.14)));
+    });
+
     test('formatDecimal', () {
       expect(formatDecimal(10.0), equals('10'));
       expect(formatDecimal(10.10), equals('10.1'));
@@ -106,6 +118,44 @@ void main() {
   group('Statistics', () {
     setUp(() {});
 
+    test('int(4)', () {
+      var data = [10, 20, 30, 40];
+      var statistics = data.statistics;
+
+      expect(statistics.length, equals(4));
+      expect(statistics.isEmpty, isFalse);
+      expect(statistics.isNotEmpty, isTrue);
+
+      expect(statistics.sum, equals(100));
+      expect(statistics.mean, equals(25));
+      expect(statistics.standardDeviation, equals(27.386127875258307));
+      expect(statistics.squaresSum, equals(3000));
+      expect(statistics.squaresMean, equals(750.0));
+
+      expect(statistics.centerIndex, equals(2));
+      expect(statistics.center, equals(30));
+
+      expect(statistics.medianLow, equals(20));
+      expect(statistics.medianHigh, equals(30));
+      expect(statistics.median, equals(25));
+
+      expect(statistics.medianLowIndex, equals(1));
+      expect(statistics.medianHighIndex, equals(2));
+
+      expect(statistics.isMeanInRange(10, 30), isTrue);
+      expect(statistics.isMeanInRange(10, 30, 25, 28), isTrue);
+
+      expect(statistics.isMeanInRange(10, 19), isFalse);
+      expect(statistics.isMeanInRange(10, 30, 20, 21), isFalse);
+
+      expect(
+          statistics.toString(), equals('{~25 +-27.3861 [10..(30)..40] #4}'));
+      expect(statistics.toString(precision: 0),
+          equals('{~25.0 +-27.386127875258307 [10..(30)..40] #4}'));
+
+      expect(data.statisticsWithData.data, equals(data));
+    });
+
     test('int(3)', () {
       var data = [10, 20, 30];
       var statistics = data.statistics;
@@ -123,6 +173,13 @@ void main() {
       expect(statistics.centerIndex, equals(1));
       expect(statistics.center, equals(20));
 
+      expect(statistics.medianLow, equals(20));
+      expect(statistics.medianHigh, equals(20));
+      expect(statistics.median, equals(20));
+
+      expect(statistics.medianLowIndex, equals(1));
+      expect(statistics.medianHighIndex, equals(1));
+
       expect(statistics.isMeanInRange(10, 30), isTrue);
       expect(statistics.isMeanInRange(10, 30, 20, 22), isTrue);
 
@@ -130,9 +187,9 @@ void main() {
       expect(statistics.isMeanInRange(10, 30, 20, 21), isFalse);
 
       expect(
-          statistics.toString(), equals('{~20 +-21.6024 [10..(20)..30] #3.0}'));
+          statistics.toString(), equals('{~20 +-21.6024 [10..(20)..30] #3}'));
       expect(statistics.toString(precision: 0),
-          equals('{~20.0 +-21.602468994692867 [10..(20)..30] #3.0}'));
+          equals('{~20.0 +-21.602468994692867 [10..(20)..30] #3}'));
 
       expect(data.statisticsWithData.data, equals(data));
     });
@@ -153,6 +210,13 @@ void main() {
 
       expect(statistics.centerIndex, equals(1));
       expect(statistics.center, equals(20));
+
+      expect(statistics.medianLow, equals(10));
+      expect(statistics.medianHigh, equals(20));
+      expect(statistics.median, equals(15));
+
+      expect(statistics.medianLowIndex, equals(0));
+      expect(statistics.medianHighIndex, equals(1));
 
       expect(data.statisticsWithData.data, equals(data));
     });
@@ -177,6 +241,13 @@ void main() {
       expect(statistics.centerIndex, equals(0));
       expect(statistics.center, equals(10));
 
+      expect(statistics.medianLow, equals(10));
+      expect(statistics.medianHigh, equals(10));
+      expect(statistics.median, equals(10));
+
+      expect(statistics.medianLowIndex, equals(0));
+      expect(statistics.medianHighIndex, equals(0));
+
       expect(data.statisticsWithData.data, equals(data));
     });
 
@@ -197,7 +268,99 @@ void main() {
       expect(statistics.centerIndex, equals(0));
       expect(statistics.center, equals(0));
 
+      expect(statistics.medianLow, equals(0));
+      expect(statistics.medianHigh, equals(0));
+      expect(statistics.median, equals(0));
+
       expect(data.statisticsWithData.data, equals(data));
+    });
+
+    test('error(alreadySortedData)', () {
+      expect(() => Statistics.compute([10, 20, 30], alreadySortedData: false),
+          isNot(throwsArgumentError));
+
+      expect(() => Statistics.compute([30, 20, 10], alreadySortedData: false),
+          isNot(throwsArgumentError));
+
+      expect(() => Statistics.compute([10, 20, 30], alreadySortedData: false),
+          isNot(throwsArgumentError));
+
+      expect(() => Statistics.compute([30, 20, 10], alreadySortedData: true),
+          throwsArgumentError);
+    });
+
+    test('multiplyBy', () {
+      var statistics1 = [10, 20, 30].statistics;
+      var statistics = statistics1.multiplyBy(3);
+
+      expect(statistics.sum, equals(statistics1.sum * 3));
+      expect(statistics.mean, equals(statistics1.mean * 3));
+      expect(statistics.standardDeviation,
+          equals(statistics1.standardDeviation * 3));
+
+      expect(statistics.squaresSum, equals(statistics1.squaresSum * 3));
+    });
+
+    test('divideBy', () {
+      var statistics1 = [10, 20, 30].statistics;
+      var statistics = statistics1.divideBy(2);
+
+      expect(statistics.sum, equals(statistics1.sum / 2));
+      expect(statistics.mean, equals(statistics1.mean / 2));
+      expect(statistics.standardDeviation,
+          equals(statistics1.standardDeviation / 2));
+
+      expect(statistics.squaresSum, equals(statistics1.squaresSum / 2));
+    });
+
+    test('sumWith', () {
+      var statistics1 = [10, 20, 30].statistics;
+      var statistics2 = [100, 200, 300].statistics;
+      var statistics = statistics1.sumWith(statistics2);
+
+      expect(statistics.sum, equals(statistics1.sum + statistics2.sum));
+      expect(statistics.mean, equals(statistics1.mean + statistics2.mean));
+      expect(
+          statistics.standardDeviation,
+          equals(
+              statistics1.standardDeviation + statistics2.standardDeviation));
+
+      expect(statistics.squaresSum,
+          equals(statistics1.squaresSum + statistics2.squaresSum));
+    });
+
+    test('statisticsMean', () {
+      var statistics1 = [10, 20, 30].statistics;
+
+      var statistics2 = statistics1.cast<int>();
+      expect(statistics2.min, isA<int>());
+
+      var statistics3 = statistics1.cast<double>();
+      expect(statistics3.min, isA<double>());
+
+      var statistics4 = statistics1.cast<num>();
+      expect(statistics4.min, isA<num>());
+    });
+
+    test('statisticsMean', () {
+      var statistics1 = [10, 20, 30].statistics;
+      var statistics2 = [100, 200, 300].statistics;
+
+      var statistics = [statistics1, statistics2].statisticsMean;
+
+      expect(statistics.sum, equals((statistics1.sum + statistics2.sum) / 2));
+
+      expect(
+          statistics.mean, equals((statistics1.mean + statistics2.mean) / 2));
+
+      expect(
+          statistics.standardDeviation,
+          equals(
+              (statistics1.standardDeviation + statistics2.standardDeviation) /
+                  2));
+
+      expect(statistics.squaresSum,
+          equals((statistics1.squaresSum + statistics2.squaresSum) / 2));
     });
 
     test('operator +', () {
@@ -246,7 +409,7 @@ void main() {
       var statistics = [10, 20, 30].statistics;
 
       expect(
-          statistics.toString(), equals('{~20 +-21.6024 [10..(20)..30] #3.0}'));
+          statistics.toString(), equals('{~20 +-21.6024 [10..(20)..30] #3}'));
     });
 
     test('DataEntry', () {